Metadata-driven dataframe casting
