+++ enki/src/test/scala/enki/StageTest.scala
@@ -5,23 +5,17 @@ import org.apache.spark.sql.types._
 
 import scala.collection.JavaConversions._
 
-// @decimalPrecision(10, 5)
-case class DecimalPrecisionTestData( a: Option[BigDecimal], b: Option[String])
+case class DecimalPrecisionTestData(@decimalPrecision(38, 12) a: Option[BigDecimal])
 
 class StageTest extends EnkiTestSuite {
   "ReadAction" should {
     "handle non-default decimal's scale and precision in strict mode" in {
 
-      val schema = StructType(Array(StructField(name = "a", dataType = DecimalType(10, 5)), StructField(name = "b", dataType = StringType)))
-      sparkSession.sqlContext.createDataFrame(Seq(Row(BigDecimal(1999), "test")), schema).write.saveAsTable("default.DecimalPrecisionTestData")
+      val schema = StructType(Array(StructField(name = "a", dataType = DecimalType(38, 12))))
 
-      sparkSession.sqlContext.table("default.DecimalPrecisionTestData").schema.foreach(println)
-
-
-      val ra = new ReadAction[DecimalPrecisionTestData](schemaName = "default", tableName = "DecimalPrecisionTestData", strict = false)
-      ra(sparkSession).schema.foreach(println)
-      println(ra(sparkSession).collect())
+      sparkSession.sqlContext.createDataFrame(Seq(Row(BigDecimal(10))), schema).write.saveAsTable("default.DecimalPrecisionTestData")
+      val ra = new ReadAction[DecimalPrecisionTestData](schemaName = "default", tableName = "DecimalPrecisionTestData", strict = true)
+      ra(sparkSession).collect() shouldBe Array(DecimalPrecisionTestData(a = BigDecimal(10)))
     }
   }
-
 }
